Necessary changes are done because time is now considered as a float type in RPMPLv2, and because of introducing 'PlannerType' in RPMPLv2.

+++ src/etf_modules/sim_bringup/include/base/BaseNode.h
@@ -26,7 +26,7 @@ namespace sim_bringup
         virtual void baseCallback() = 0;
 
         rclcpp::TimerBase::SharedPtr timer;
-        int period;                             // Period of basic callback function in [ms]
+        float period;                             // Period of basic callback function in [s]
 
     protected:
         std::string project_abs_path;
